All fixes are completed
