+++ src/data/directions.ts
@@ -5,7 +5,7 @@ const waterCheckpoints: Checkpoint[] = [Checkpoint.NorthernWater, Checkpoint.Sou
 const earthCheckpoints: Checkpoint[] = [Checkpoint.Abbey, Checkpoint.GaipanVillage, Checkpoint.SiWong, Checkpoint.BaSingSe];
 const fireCheckpoints:  Checkpoint[] = [Checkpoint.FireCapital, Checkpoint.ShuJing]
 
-const nationByCheckpoint: {[key: CheckPoint]: NationType} = {
+const nationByCheckpoint: {[key: Checkpoint]: NationType} = {
     [Checkpoint.NorthernAir]: NationType.Air,
     [Checkpoint.SouthernAir]: NationType.Air,
     [Checkpoint.EasternAir]: NationType.Air,
