+++ src/views/HomeBisonView.vue
@@ -1,15 +1,14 @@
 <script setup lang="ts">
 import type {Carriage, PackageState, User} from "@/types/intefaces";
 import {getCurrentUser} from "@/services/user";
-import {carriagesByBison, getCurrentCarriage} from "@/services/carriage";
-import {getCurrentPackage, packageByBison} from "@/services/package";
+import {getCurrentCarriage} from "@/services/carriage";
+import {getCurrentPackage} from "@/services/package";
 import HeaderName from "@/components/core/HeaderName.vue";
 import Hero from "@/components/core/Hero.vue";
 import BisonOrder from "@/components/bison/BisonOrder.vue";
+import NoOrders from "@/components/bison/NoOrders.vue";
 
 const user: User = getCurrentUser()
-const myCarriages: Carriage[] = carriagesByBison(user?.id!)
-const myPackages: PackageState[] = packageByBison(user?.id!)
 const currentPackage: PackageState | undefined = getCurrentPackage(user?.id!)
 const currentCarriage: Carriage | undefined = getCurrentCarriage(user?.id!)
 
@@ -23,25 +22,18 @@ const order: Carriage | PackageState | undefined = currentPackage? currentPackag
     name: user.name,
     message: 'Bienvenido a la mejor aplicación de pedidos y acarreos.'
     }"/>
-    <h1>Acarreos</h1>
-    <h1 v-for="carriage in myCarriages">
-      <br>
-      <a href="">{{ carriage }}</a>
-    </h1>
-    <br>
-    <h1>Paquetes</h1>
-    <h1 v-for="pkg in myPackages">
-      <br>
-      <a href="">{{ pkg }}</a>
-    </h1>
     <BisonOrder v-if="order" :order="order"/>
-    <h1 v-else>No tienes pedidos asociados</h1>
   </div>
+    <NoOrders class="mt-5" v-if="!order"/>
 </template>
 <style scoped>
 .home-page {
   width: 92%;
   max-width: 80rem;
   margin: 0 auto;
+}
+
+.no-order {
+  margin-top: 30px;
 }
 </style>
